requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
click==8.1.7
httpx==0.25.1
web3==6.11.1
pydantic-settings==2.1.0
python-dotenv==1.0.0
nats-py==2.4.0
